requirements.txt
Fabric==1.4.3
pycrypto==2.6
ssh==1.7.14
wsgiref==0.1.2
Jinja2==2.6
flask==0.9
Werkzeug==0.8.3
envoy==0.0.2
slimit==0.7.4
cssmin==0.1.4
poster==0.8.1
boto==2.6.0
requests==0.14.1
python-tumblpy==0.6.3
python-tumblr==0.1
